requirements.txt
fastapi
uvicorn
python-dotenv
twilio
elevenlabs
websockets